fixing react render warning
