+++ droidlet/dashboard/web/src/components/Timeline/TimelineUtils.js
@@ -2,6 +2,8 @@
 Copyright (c) Facebook, Inc. and its affiliates.
 */
 
+import Fuse from "fuse.js";
+
 export function capitalizeEvent(str) {
   // replaces underscores with spaces
   str = str.replace(/_/g, " ");
@@ -47,4 +49,39 @@ export function handleClick(stateManager, item) {
   stateManager.dashboardLayout.root
     .getItemsById("timelineDetails")[0]
     .addChild(config);
+}
+
+export function handleSearch(stateManager, pattern) {
+  const matches = [];
+  if (pattern) {
+    stateManager.memory.timelineSearchPattern = pattern;
+    const fuseOptions = {
+      // set ignoreLocation to true or else it searches the first 60 characters by default
+      ignoreLocation: true,
+      useExtendedSearch: true,
+    };
+
+    const fuse = new Fuse(
+      stateManager.memory.timelineEventHistory,
+      fuseOptions
+    );
+
+    // prepending Fuse operator to search for results that include the pattern
+    const result = fuse.search("'" + pattern);
+
+    if (result.length) {
+      result.forEach(({ item }) => {
+        const eventObj = JSON.parse(item);
+        if (
+          stateManager.memory.timelineFilters.includes(
+            capitalizeEvent(eventObj["name"])
+          )
+        ) {
+          matches.push(eventObj);
+        }
+      });
+    }
+  }
+  stateManager.memory.timelineSearchResults = matches;
+  stateManager.updateTimelineResults();
 }
